refactored view to separate UI creation from logical data instantiation UI created with injection
